RSE: Allow RSE to use local UART
Allow the TC platform in RSS to use the UART local to the RSS subsystem.
We achieve this by adding a new RSS_USE_LOCAL_UART flag. In order to
make use of this platforms must define the secure and non-secure UART base
addresses using LOCAL_UART0_BASE_S and LOCAL_UART0_BASE_NS respectively.

Note we also need to add a TC specific tfm_peripherals_def.c file to
define the tfm_peripheral_std_uart structure should we be using the
local UART. This is added to the common directory and included in
the target sources for each TC platform. The reason this is local to TC
is to avoid unnecessarily modifying common code and also the UART
definition is local to the TC platform so it does not make sense to
expose this address outside of the TC directory.

Change-Id: I65e24abeaa9db308da0a680b96b1bee0a42de43b
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>