Change fih_eq() and fih_not_eq() to macros
The resulting code is more resistant to injected faults.

This has been tested with a simulator that advances the PC to "skip"
instructions.

This reflects the change adopted in mcuboot in
https://github.com/mcu-tools/mcuboot/commit/5047f032c93fd8957be6529f8f1ad7a3fe626057
This version, though, retains the two calls to fih_delay() from
the previous TF-M implementation of the double variable version.

Signed-off-by: Chris Brand <chris.brand@cypress.com>
Change-Id: Ic95b10c381a9eff061f42241dd519282f561321e