rse: rd1ae: Wait for SCP signal before AP BL2 load
Before MCUBoot can load AP BL2 firmware, the SYSTOP power domain needs to
be in ON state. The SYSTOP power domain is switched on by SCP, so RSE
has to wait for SCP to signal when the SYSTOP init is complete, before
MCUBoot can load the AP BL2 firmware.

When SCP has initialized SYSTOP, it sends an MHU message to RSE via the
SCP to RSE MHU. Hence, configure RSE to wait for this message before
letting MCUBoot try to load the AP BL2 firmware image.

In order to monitor the status of the host systems such as SCP and AP,
as well as to orchestrate the comms between RSE and host systems,
introduce host system as abstraction layer. Configure the SCP to RSE MHU
interrupt to notify host system module when the SYSTOP ready signal is
received so host system can be queried about whether SYSTOP is ready.

Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>
Change-Id: Ief324729bac1f48906a1304e789b3289e8b6fc7a