platform: nordic_nrf: Remove flash driver guard
Don't guard the Flash driver with the ITS since it is
used from more users (for example default OTP counters).

The RRAM and NVMC drivers are already included in the build
without any conditions so guarding the Flash driver doesn't
add any value.

This fixes a build issue with unresolved Flash driver symbols.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Change-Id: I639f7aae89eeaa24d8b6b0c0d236bd9a39dc95e5