rse: rd1ae: Add rse_scp_axis PSAM config
Add PSAM address map to map access from RSE and SCP targeting
Cluster Utility address space. All access from RSE and SCP
that target to Cluster Utility is routed via a NIC400 outside
of System Control NI-Tower and sent to the 'rse_scp_axis'
on the NI-Tower.

This is necessitated because address filtering on transactions
based on the Cluster Utility memory map is required in the NI-Tower,
but the target 'clus_util_axim' interface APU can only do address
filtering for upto 8 Clusters.

Hence, for every transaction targeting the Cluster Utility, the upper
address bits are propagated via AXI USER bits rather than on the AXI
address signals and the address filtering is applied based on the
lower address bits. With this, NI-Tower has to handle only on a 'single'
Cluster in the system, regardless of the number of Clusters in the
system or regardless of the chip address offset.

Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>
Change-Id: I5b1761b664708d492ca2e9ea1f8bb4eb68089c58