Crypto: psa_get_key_attributes() must not overwrite the out_vec
psa_get_key_attributes() has psa_key_attributes_t as an output
parameter but the size of it in the client view (i.e. the caller)
will be always 4 bytes short respect to the size of it as being
processed by the Crypto service. Make sure that a service key
attributes structure is passed to the backend and copy back only
the fields that are available in the client side before finishing.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: I8a846b72cee55044c501c11885ec6d57412d13ad