SPM: Panic if metadata is broken
As the code stands, we check for NULL before updating the global
metadata pointer, but do nothing if the check fails. That could leave us
with the metadata for the previous partition, but the PSA API functions
are part of that metadata, so having the wrong ones in place means that
TF-M is broken.

So let's panic cleanly as soon as we detect that.

Signed-off-by: Chris Brand <chris.brand@cypress.com>
Change-Id: I890bf5b07ab98f413695b76320b07c1158f452a8