iar: Fix v8.1m arch check
IAR has a different architecture defines than GCC and Armclang.
CMSIS headers have these definitions, but you always have to
include the cmsis_compilers.h file to activate it, but nothing forces it.
To have these definitions in all cases, I've added the global
defines for every supported architectures.

The CMSIS patch is upstreamed to the CMSIS6 repo as a PR.

Signed-off-by: Dávid Házi <david.hazi@arm.com>
Change-Id: Iaaee265f77f23507168df9b46884d0b7d3818bea