Build: Add support for PACBTI
PACBTI (Pointer Authentication and Branch Target Identification) is an
optional feature to improve robustness of the system, preventing some
attacks like Return-Oriented Programming (ROP) and Jump-Oriented
Programming (JOP).

This feature needs to be enabled by the compiler and needs
architectural support (Armv8.1-M).

This patch adds support for PACBTI in the build system for GNU and
ARMClang compilers. It is provided for the SPE build only.

It is by default DISABLED.

To enable the feature, use the CONFIG_TFM_BRANCH_PROTECTION_FEAT option.

Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Change-Id: I0a3542501ce040a86a58f1bd3b71ab48f4e041eb