RSE: Add BL1 test framework
Add extra testsuites for BL1_1 and BL1_2. Add framework to allow tests
to be run in only certain LCSes or SP modes, and add state transitions
tests so that every LCS/SP mode is covered. A random PCI/TCI mode is
also chosen for the tests, in order to cover both as we can't switch TP
mode during a single test run.

Change-Id: Ia60e04c95c926018839477fd6010ea123b27e7a7
Signed-off-by: Raef Coles <raef.coles@arm.com>