SPM: Remove specific section for psa_interface_thread_fn_call
A previous commit
cbcd7877d ("SPM: Put psa interface thread call into one section")
added a "__section" attribute to the functions contained in
psa_interface_thread.c.
This was done because at the time the armclang compiler
generated errors on Arm v6m/v7m.

It appears now that when building for GCC, those functions are not
placed within the .text section.

Remove the explicit attribute such that the code is placed in
the .text section as expected.

Tested successfully with the following compilers/versions:
 - Arm Compiler for Embedded 6.19 Tool: armlink [5e73cb00]
 - Arm Compiler for Embedded 6.18 Tool: armlink [5e4cc000]
 - ARM Compiler 6.14.1 Tool: armlink [5db06d00]
 - ARM Compiler 6.14 Tool: armlink [5db06800]
 - ARM Compiler 6.13 Tool: armlink [5d895400]
 - IAR ARMBX.EW 9.50.
for both AN521 and PSOC64.

In all the above cases the functions are placed within the .text
section.

Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Change-Id: I4d967012e3e6d65a9b4a79a2ceee1c7cc54d11de