RSE: RD-V3: Fix missing APU enable during post load stage of AP-BL1
AP-BL1 region should remain read-only as it is considered equivalent to
ROM for the AP boot. The program_periph_ni_tower_post_ap_bl1_load()
function is used to make the AP-BL1 region as read-only by initializing
the AP-BL1 regions in the APU but enabling the APU was missing. Add
ni_tower_apu_enable call to enable the APU otherwise the AP-BL1 region
remain RW.

Signed-off-by: Arnold Gabriel Benedict <arnoldgabriel.benedict@arm.com>
Change-Id: Icc9109f0988f8bdfaaac96ca862d6cd52b5fe1cc