CC3XX: Add driver testsuite
Add the development testsuite for the CC3XX driver into the driver
directory, for use in further development and for validation of CC3XX
hardware integration. Test cases are not complete, but cover a
significant part of the driver functionality. Note that because this is
an internal testsuite converted into a TF-M testsuite, the the grouping
of the testcases is non-uniform, mostly because it's not possible to
pass arguments into a tfm_test_t.

Change-Id: I860edf66447b55715046e7bdcc0b21777e692666
Signed-off-by: Raef Coles <raef.coles@arm.com>