CC3XX: Reset aes_remaining_bytes before final GCM encryption
When performing an authentication only GCM operation, the value
of aes_remaining_bytes is set just before the a call to the DMA
flush. But if the engine has never been set for crypted data,
the AES engine is not called hence the value which was set for
aes_remaining_bytes is not decremented. Make sure that is cleared
before the final encryption so that the proper tag length is produced.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: I1e75a67fa5547531fcbc858a4daaa99e2198062e