Build: Suppress armclang and IAR linker error
There are symbols in the bl1_1_shared_symbols.txt that are
only defined in some configuration. If other configurations are
used then armclang and IAR throw error for missing definitions.

This error is suppressed by this commit so the same shared symbol
file can be used for every configuration, without conditions.

Change-Id: I9e35ac1105a02d6b345b1e06f863ca4fc31e9f2e
Signed-off-by: Bence Balogh <bence.balogh@arm.com>
Signed-off-by: Dávid Házi david.hazi@arm.com