SPM: Combine fixes for reverse-handle issue
*** This is a squash of previous three commits [1]:

e9fb54f ("SPM: Use p_replied handles to manage reply message")
362e318 ("SPM: Remove is_tfm_rpc_msg()")
5f3cae4 ("SPM: Avoid overriding the message rhandle in the backend")

Note that the above SHA are not in tree, but in gerrit only.

The squash has become necessary to:
- successfully pass the tests in CI (the rhandle test has been introduced
  while these changes were refined)
- avoid sparse cherry-picking leaving behind required changes to resolve
  the rhandle issue
This commits comes with a new Change-Id to preserve the history
in gerrit for the original patches.

Credits go to the original author:
Jianliang Shen <jianliang.shen@arm.com>

Here's a summary of the changes:

---- 1 ----
SPM: Use p_replied handles to manage reply message

The p_handles in partition is used to record the request connections
of the service. The Mailbox NS Agent uses it to collect the handles
ready to be replied. It works as current agent partition doesn't
have any services.

Create the replied handles list mounted in the partition to manage
reply message in the IPC model. Both synchronous and asynchronous
replies can be handled by their own reply implementation:

  - Synchronous reply will change the block state of partition thread
    and take away the return value from the only handle to be replied
    in the list.
  - Asychronous reply is supported for Mailbox NS Agent currently.
    It takes away the handles from the list and reply the message
    one by one.

---- 2 ----
SPM: Remove is_tfm_rpc_msg()

The Mailbox NS Agent doesn't have services. It is assumed
to be the client now. IS_NS_AGENT_MAILBOX() can help
check the connection from the RPC as well.

---- 3 ----
SPM: Avoid overriding the message rhandle in the backend

The connection handle of Mailbox NS Agent stores the
client data of the mailbox message handle which is used
for mailbox message reply. Transmit the connection handle
to the message 'rhandle' before replying to avoid overriding
the other potential value used in the backend.

The reply of Mailbox NS Agent is implemented by getting
the message with ASYNC_MSG_REPLY signal. Create a specific
function to distinguish it from common routine.

[1]
https://review.trustedfirmware.org/c/TF-M/trusted-firmware-m/+/25386
https://review.trustedfirmware.org/c/TF-M/trusted-firmware-m/+/25384
https://review.trustedfirmware.org/c/TF-M/trusted-firmware-m/+/25239

Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Change-Id: I07a25efd6385992873df0000a5ce7bf4a83a6bcf