Lib: t_cose: Change library sources to private
Changes the tfm_t_cose_s library so that only the t_cose headers and
defs are public, with the rest of the library private. This ensures
that the library sources are part of the library in TFM_UNPRIV_CODE
region, rather than ending up exported into partitions. This fixes the
DPE partition secure tests in isolation level 2.

Add link to psa_crypto_config to initial attestation partition, which
was relying on indirect link via tfm_t_cose_s.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: I48f3511e27f648daf09a2ed01ad3a59272ef3a6a