TFM: Explicit use of SPM_ASSERT only for secure code
SPM_ASSERT is available to assist secure code to catch unintended
behaviour and to print the location of the failure.

Platform code can use such macro but it would be available only when
building for secure code. This caused build issues when SPM_ASSERT is
used in non-secure platform code or platform code ised in BL1/BL2
stages.

Firstly, introduce a distinct header for SPM_ASSERT in a private include
directory, such that its inclusion is made explicit when the macro is
used.

Secondly, supply a transparent mechanism where assert() in platform
code expand to either:
- SPM_ASSERT(), for runtime firmware only
- standard library asserts, for non-secure code or BL1/BL2 code

Lastly, explicitly include the new private assert header where
required. Note that for platform code this is temporarily and will
be removed in the upcoming changes.

Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Co-Authored-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: If38136d9327b85bb9dc1361ecc8f3f777f74b3d3