rse: rd1ae: Move AP BL2 to AP Secure Flash
The Application Processor BL2 is now part of the FIP image in
the Application Processor Secure Flash. Modify TF-M BL2 to extract
the AP BL2 from the FIP image.

Note: The flash map structure needs to be changed to non constant
as it now gets modified during runtime to point to the correct
address of AP BL2 in the FIP.

Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>
Change-Id: Ib613d54ec57a722bcbe41490e52ccdc1b616b36f