Build: Remove unused ext/driver includes
Following change 1ec133d24 it is no longer necessary to install the
platform/ext/driver directory to NS, nor to include it in the NS build
or flash libraries. This patch removes some leftover cases that were
missed by 1ec133d24 and adds a link to the cmsis library where needed
to fix some build configs.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: I387ac79b594401f464762027e160d5e0965915bb