CC3XX: Do not count the output of ccm_input_auth_length()
The input of ccm_input_auth_length() currently is added up
to the aes_state.authed_length counter, leading to a situation
where this counter at the end of the AD update is greater than
the expected data to be authenticated. Although it does not
have a practical impact at this stage, it should be consistent.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: Id025bf2379c6ea828d4d652bd6b105fbce975bec