build: remove the warning when the version Git tag is not found
In the case of Zephyr's integration of TF-M, the Git tags of the
upstream TF-M repo are not fetched.
This resulted in this warning firing up at every build for every user
while in fact it does not require any action from the user, and is
just noise.
Thus remove it.

Change-Id: I9fc78ff89e978a0622e80d59dadc8dcfeeb7e553
Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>