[zep noup] platform: ext: add local copy of CMSIS v6 files
This adds all the files (minus `.git*` and `CMSIS/Documentation/`
for saving on size) from the CMSIS v6 repository
(https://github.com/ARM-software/CMSIS_6) at the revision `d0c460c169`
as defined in `lib/ext/cmsis/CMakeLists.txt`.
The patch `lib/ext/cmsis/0001-iar-Add-missing-v8.1m-check` is applied
on top.

This is because as of v2.1.0 TF-M has updated to CMSIS v6 and switched
from hosting the sources to depending on the upstream repository,
cloning it at build time.

To prevent a download from happening during the build, CMSIS v6 sources
are pushed and the CMSIS_PATH CMake variable is used to point to them.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>