PS: Make PS crypto less stateful
By constructing the label for the object table from a dummy client_id
and UID, as is done for objects, the label construction and key
generation can be moved entirely inside ps_crypto_interface, which
reduces the amount of state stored there and removes some file-scope
variables.

Note that this change alters the format of the persistent PS data, so
existing saved data from versions of TF-M prior to this patch will need
to be migrated to the new format or re-created from scratch.

Signed-off-by: Chris Brand <chris.brand@cypress.com>
Change-Id: I7e034fd3abf7e677449d7148972f4f3531155d88