rse: rd1ae: Fix GIC 720AE driver and configurations
This change includes the following fixes for the GIC 720AE
driver and its configurations:

1- The size of the Safety Island GIC View should be 0x400000
instead of 0x200000.
2- The FMU CI block is powered on only when the GIC View 0
redistributor is powered on, so implement the controls
to do that through the GICR_PWRR register.
3- Enable all of the GIC 720AE redistributors from RSE
since the GICR_PWRR registers are only accessible from View 0.
4- The GICD_IVIEWRn registers start at 0, with the first 2 registers
reserved. There are 960 SPIs controlled by those registers,
so fix the computation of the SPI_IS_VALID macro as well.

Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>
Change-Id: Icb82c437d5f87ab03ae2ef2e19cfc33d8431b557