Mailbox: Optionally cache align NS mailbox struct
The NS mailbox is used to communicate between the SPE and NSPE cores.
Either or both may have a dcache, and the mailbox itself may or may not
be in cacheable memory.

In case the NS mailbox may be cached on one or both sides, introduce
cache alignment where appropriate. The alignment changes can be disabled
by platforms, as they will enlarge the mailbox itself somewhat. This
change means that cache clean and invalidate instructions can be
introduced without erroneously invalidating data that has been changed in
the cache but not yet written to memory.

Signed-off-by: Chris Brand <chris.brand@cypress.com>
Change-Id: I65162684b6a71e65e934530f4bd0293bff6ebddd