RSE: avoid overflow and remove always false branch
`tfm_plat_increment_nv_counter` performs a check for the security
counter being equal to `UINT32_MAX` which depending on the OTP
configuration would either correspond to `UINT32_MAX` bits or OTP words.

For all practical purposes the check would return false. A better check
can be performed in `read_otp_counter`/`set_otp_counter` depending on
the OTP configuration by comparing the set bits or words with the
bits/ words configured in the otp layout.

`set_otp_counter` fails with an overflow in case the counter size
configured in the otp layout is less than `OTP_COUNTER_MAX_SIZE` since
`tfm_plat_otp_write` is called with the maximum size instead of the
size configured in the otp layout.

Change-Id: Ib69fe98001a1ea5e5f6f4c65349600a67d1c5d37
Signed-off-by: Anubhav Raina <anubhav.raina@arm.com>