RSE: TC: Allow use of SCP CPUWAIT register
TC platforms use two mechanisms to release the SCP from reset, depending
on the hardware. Either use the RSE retention register or the SCP
cpu wait register. This patch adds a new config flag which specifies
use of the SCP register instead of the RSE one.

By default, this is enabled for TC4 and disabled for TC3, as is
consistent with the most commonly used hardware.

Change-Id: I701703dd97ebab20105c0ce42bea8df0ad62fd81
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>