RSE: Set default MCUBOOT_HW_KEY for BL2
The enablement patch of EC-P256 has also enabled the
MCUBOOT_BUILTIN_KEY switching off MCUBOOT_HW_KEY. In
this way, the entire public key must be provisioned
in the OTP memory for RSE instead of just the hash.
This becomes a trade off that each platform needs to
make, but it tends to diverge quite a lot in size
especially if multiple images have to be loaded. For
this reason, revert the default back to HW_KEY and
leave platforms to decide based on their own tradeoffs.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: I8ad5f3078ee446b830f8ede0519d16a087ab86fd