Platform: Switch multi-core platforms to new linker files
Updates the multi-core mem check implementation to use the PT_<region>
position tags of the new combined isolation L1-3 linker files to
determine the attributes of each memory region. Changes PSoC 64 and
Corstone-1000 to use the new linker files.

Adds a separate alignment definition for the individual partitions'
private data regions, so that Corstone-1000 can align the boundaries
of the whole App RoT data region without introducing excessive padding
by aligning each partition region within it.

Changes the linker script to only relocate the BL2 shared data area if
S_DATA_PRIV_START is defined, and undefs it for Corstone-1000. On this
platform, BL2 was already writing shared data at the beginning of SRAM,
which previously overlapped with uninitialized heap memory at runtime,
and so the issue wasn't noticeable. As Corstone-1000 wasn't relying on
S_DATA_PRIV_START to configure its MPU regions, the simplest solution is
to undefine it to move the BL2 shared data link region in the runtime to
the position where the data is currently being written.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: I34dcfb322063e67c88b43c94da89c105493df4a0