RSE: TC: Reset SDS region on BL2 init
This region is used by the RSE and AP for DPE related synchronisation on
startup. On reset, the memory region is not reset to 0 so the AP will
wrongly assume that the RSE NS has stared. Memset the region to 0 to
prevent this and cause the AP to spin on the SDS region's initialisation
as with the initial boot.

Note that we make use of the same temporary ATU region as has been
previously defined for boot_hal_bl2.c.

Change-Id: I23096157dd67603b6f1a3e8cd53e7973f3282f45
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>