SPM: mailbox_agent_api: Free connection if params association fails
After the connection is successfully allocated, if the parameters
association fails, then the connection needs to be released.
This is only required for STATELESS connections for the reason that
for stateful connections those are taken care of as part of the
psa_close sequence.

Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Change-Id: Ic0674098b7780a4e83b21fe93c5ed83ff5a2e8d1
(cherry picked from commit 417063d18e7004f9a50ddeb9e6ce78441eaa46b2)