Platform: CS1000: Increase NS mailbox stack size
The Mailbox partition is used as an NS Agent so its stack size is used
to determine the Process Stack bottom and PSLIM (Process Stack Pointer
Limit) during the SFN (Secure Function) backend initialization.

It has to be increased because a stackoverflow could happen with the
default 0x800 value which is coming from the TF-M default config files.
The new 0x1000 was chosen by trial, the regression tests pass with
this stack size.

Change-Id: Ib3fd523c4a1c3cf30d520eb0cb00c130c684eed5
Signed-off-by: Bence Balogh <bence.balogh@arm.com>