CC312: ADAC: Add PSA_WANT_ALG_SHA_256 definition
The bl2_mbedcrypto_config is linked to the psa_adac_cc312 target
so the MCUBOOT_PSA_CRYPTO_CONFIG_FILEPATH and
MCUBOOT_MBEDCRYPTO_CONFIG_FILEPATH configs are used for the ADAC
driver too. If MCUBOOT_USE_PSA_CRYPTO is OFF, that
means the MCUBOOT_PSA_CRYPTO_CONFIG_FILEPATH is not included
during the build so the PSA_WANT_ALG_SHA_256 is not defined for
the ADAC driver. Because of this, the PSA_HASH_MAX_SIZE is not
set correctly for the sources of the psa_adac_cc312 target. This
caused runtime issues. Note that this is relevant only when
MCUBOOT_USE_PSA_CRYPTO is not defined, as in that case the config
file is correctly included and propagated to the target.

Signed-off-by: Bence Balogh <bence.balogh@arm.com>
Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: I50d69dc96f0ce7e66d55492b0c71105e650d664d