Platform: RSE: Add volatile qualifier to local buf
The integrity check value can be copied to a local buffer in the driver
before being checked, or read out from a local buffer after being
generated. At -O3, the compiler deduces that the local buffer is written
but not read (by the CPU), or read without being modified, and
eliminates the write or read. This patch adds a volatile qualifier to
the buffer to ensure the reads & writes actually occur.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: Icffb40ee1701079171cd0668c220b5b91e7fcd18