RSE: Fix BL2 region size error
There were two issues with the CPP conditional which determined whether
or not we are overflowing the SRAM:
- XIP_TABLE_SIZE is not defined (will default to 0). This macro should
be BL2_XIP_TABLES_SIZE
- FLASH_BL2_PARTITION_SIZE + BL2_DATA_SIZE + BL2_XIP_TABLE_SIZE will
always add up to VM0_SIZE + VM1_SIZE (as that is the way that
BL2_DATA_SIZE is defined) and therefore the error will never be caught.

Instead change this CPP conditional to determine if the BL2_DATA_SIZE is
greater than 0. This ensures that BL2_XIP_TABLES_SIZE +
FLASH_BL2_PARTITION_SIZE is never greater than VM0_SIZE + VM1_SIZE.

Change-Id: I528d86a37cdb62e2769a41d8a209a3f8eb48aa13
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>