RSE: TC: Increase TC3/4  BL2 code partition size
Increase the size of the FLASH BL2 partition size as the TC3/4 Debug
builds were overflowing the BL2 FLASH region. Increasing value of this
macro results in the RAM portion being allocated to the code increasing
in size by reducing the size of the data portion. This is ok for TC3/4
builds as the new data region is still large enough.

Change-Id: I43805af1e635d9e8156711878c93af2feb7b3e99
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>