Platform: RSE: De-duplicate MHU device definitons
RSE has up to 9 CMU MHUs defined in the base architecture. Different
platforms may configure RSE with different numbers of MHUs and connect
them to different endpoints, but the MHU devices themselves are the same
if present. Therefore this patch moves all 9 MHU device definitions to
the RSE common code, and keeps only the configuration of which MHUs to
enable and their user-friendly name in platform-specific code.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: I3347d4d628411aba1dd858137ca9f47898ff60a7