CMake: Append SHA for TFM_VERSION_MANUAL
When the 'TF-M' string is not found in the git describe
reference, TFM_VERSION_MANUAL is then used.
This is the case when the latest tag is not TF-Mv2.X.Y,
but a custom tag.
When this happens, the SHA is not available in the final
version, so it makes difficult to pin exactly the version
or change of the software used.

Thus, append the SHA even when the TFM_VERSION_MANUAL is used.

With this change, the f/w boot message shown at boot will
(currently) be: "Booting TF-M 2.1.0+g<SHA>".

Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Change-Id: I8c0c807d2d9983a7444875910b71fd0ae7f343c9