SPM: Set DIT bit for Armv8.1-M architecture in S world
If the underlying implementation can't guarantee the DIT
behaviour, the write is ignored. TF-M threat model does not
aim to protect against side channel attacks led through
timing analysis, hence this addition is not mandatory to
be implemented for a valid implementation, but if available,
then it is good to be used.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: If88790ff2c568d0e8879eb981d780c61943eee36