build: Remove forced NS init and startup files
Startup, init and some driver files were forced during NS build. This
default behaviour can now be configured with the PLATFORM_CUSTOM_NS_FILES
variable. The compile definitions previously part of the platform_ns
library are required for security, these can not be disabled.
The startup and init files are now linked to platform_ns instead of
tfm_ns. Previously this caused weak symbols to not be correctly
overridden. This issue did not occur this time.

Signed-off-by: Gergely Kovacs <Gergely.Kovacs2@arm.com>
Change-Id: I2dc2f62cd5fd9eeb72f7b6cac863513bdb483a27