Mailbox: Remove alignment requirement on the status pointer
This allows to build with IAR as well, because IAR can't have
stricter alignment on the stack (auto variables) than the stack
alignment itself (8 bytes).

This structure is only written once in the initialization
phase by the NSPE and only read once by the SPE. Caches
are not maintained during this transaction.

Signed-off-by: Dávid Házi <david.hazi@arm.com>
Change-Id: I7c07fa98582785b3415f3e41f937254fce2b01cd