RSE: TC: Commonize header files
With the deprecation of TC2, TC header files are now the same across TC3
and TC4 and can therefore be commonized. This commit move the following
headers to the common TC directory:
- host_base_address.h
- flash_layout.h
- rse_memory_size.h

Note that config_tfm_target.h is not completely commonized because the
file is expected to live in the platform directory by the build system.
We do, however, add a new common version of this file which is included
by the platform specific files.

Change-Id: I2a35e758f0a3878e5ff78eb2409f2a7276412a01
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>