FWU: Make platform specific TFM_FWU_BOOTLOADER_LIB selectable to add
Prepare the environment where partition firmware update can be
enabled and platform specific bootloader configuration can be used.

FWU implementation provides an abstraction for the bootloader.
This bootloader abstraction layer is implemented for MCUBoot.
It can be used after making changes that can handle platform
specific behaviors. But the implementation limits it.

When TFM_PARTITION_FIRMWARE_UPDATE is enabled, the configuration
becomes invalid. Therefore, the invalid configuration is limited
to the case where TFM_FWU_BOOTLOADER_LIB is used for MCUboot.
This makes the configuration valid when a platform specific
configuration is used.

TFM_FWU_BOOTLOADER_LIB can only be added from a subdirectory, which
prevents the use of platform-specific bootloader configurations.
The logic has been changed to allow the use of platform-specific
bootloader configurations.

Change-Id: I4b258ca09f3f90580f2bf2f5b052c8c65eae5478
Signed-off-by: Ali Can Ozaslan <ali.oezaslan@arm.com>