Platform: corstone1000: Permit FWU calls in RSE-COMMS
Allow FWU calls to be dispatched by RSE-COMMS for Corstone-1000.
This change is required to allow the transmission of PSA FWU related
calls between Cortex A and Cortex M side on Corstone-1000.
For every PSA call from A side, the RSE-COMMS at M side validates, if the
call is allowed or not.

Change-Id: Ib87fd4a5133f9bfad9ef5fa9c2febf3f88e9ee62
Signed-off-by: Harsimran Singh Tungal <harsimransingh.tungal@arm.com>