mbedtls: align crypto_sizes.h to the Mbed TLS one
crypto_sizes.h is no more aligned to the Mbed TLS version and
it misses some #define that make some test to fail at build
time in Zephyr. This commit fixes this disparities.

Note: this commit can be ignored in future TF-M repo updates
assuming that this fix will be already included in the upstream
version of TF-M.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>