npu: Update NPU driver to version 24.08
In the newer version the device is not allocated
dynamically, so we can remove the patches too.

Cherry picked from 75c0999bd55b64b260529651e7ca994989508e1b
Note that the corstone320 changes from the original commit are
skipped from the cherry pick since the support for corstone320
is not yet available in Zephyr.

Signed-off-by: Dávid Házi <david.hazi@arm.com>
Change-Id: I6bbefa490f85bfa1884587fbcce3c83d11ae15aa
Signed-off-by: Sudan Landge <sudan.landge@arm.com>