SPM: psa_api: mapped and unmapped are mutually exclusive
The field iovec_status is used to track the status of memory-mapped invecs
and outvecs through an encoding. Two bits used respectively for mapped and
unmapped status are set but not individually cleared.

In some places, those bits are checked but their conditions do not
reflect the right status for the vectors.

Improve the intent and use of the encoding by making _MAPPED_BIT and
_UNMAPPED_BIT mutually exclusive.

This avoids the situation where a mm-iovec is first mapped and then
correctly unmapped, while the respective status is both mapped and unmapped,
making checks confusing and not valid.

As a consequence, update the reporting of 'no data written' in outvects for
update_caller_outvec_len.

Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Change-Id: I3137e74bc2d29b296ad698a41ee7896ab90f2110