Crypto: crypto_init: Fix missing unmapping of input vectors
tfm_crypto_call_srv performs the initialization and mapping of mm-iovecs
and after the required operation is completed, the outvecs are unmapped.

The invec are not unmapped though, resulting in a situation where
subsequent PSA calls would fail.

Note that currently this is not immediately visible due to the fact that
PSA crypto services are accessed via a static handle, so each call takes
a new connection with a clean mm-iovec status.
It would take a stateful Crypto service to actually fall into the issue
noticed above.

Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Reported-by: Brian Quach <brian@ti.com>
Change-Id: Idb4dc40ff439e177c64ddad2421f8febf6007dcb