RSE: TC: Add common host_device_cfg.h
In addition to the previous changes making TC headers common,a common
host_device_cfg.h can also be added. Note that we do not completely move
this header to the common directory, as the MHU device used for
SCP<->RSE communication differs between TC3 and TC4. Therefore, we need
to keep some definitions in the platform specific directory.

Change-Id: I564b5aa84ab1fd9c59738ce6e9d14d927f8abb37
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>